+++ lib/angular.dart
@@ -1,7 +1,7 @@
-import 'package:angular/angular.dart';
+//import 'package:angular/angular.dart';
 import 'src/generic.dart';
 
-abstract class DisposableComponent extends Disposable implements OnDestroy {
+/*abstract class DisposableComponent extends Disposable implements OnDestroy {
   @override
   void ngOnDestroy() => dispose();
-}
+}*/