+++ CHANGELOG.md
@@ -6,3 +6,7 @@ First try of the package with the wrappers:
 
 # 0.0.3
 Removed the NNBD stuff for now.
+
+# 0.0.4
+Added the disposable method. 
+It enables adding disposable objects to be disposed within another Disposable object.